logo + image fix and component for time formatting
