+++ client/src/utils/FormattedDate.js
@@ -1,0 +1,13 @@
+import React from 'react';
+const options = {
+  year: 'numeric',
+  month: 'long',
+  day: 'numeric',
+  hour: 'numeric',
+  minute: 'numeric',
+  second: 'numeric'
+};
+
+export default function (props) {
+  return <>{new Date(props.date).toLocaleDateString(undefined, options)}</>;
+}
